src/evolvefg/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-extensions==3.1.1
django-shell-plus==1.1.7
djangorestframework==3.12.2
django-cors-headers==3.11.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
mariadb==1.0.11
requests
configparser
evolved5g
pyJWT==1.7.1